Added a few mutant heads and a bugfix regarding finding food freezing creatures if they detected food outside of their reach.

+++ scripts/reproduce/reproduce.gml
@@ -6,7 +6,7 @@ var species = argument0;
 var parent = ds_list_find_value(species.creatures, random(ds_list_size(species.creatures) - 0.5))
 var parentExists = pointer_null;
 
-if (parent != pointer_null) { //Make sure the parent exists and is not null
+if (is_undefined(parent) == false) and (parent != pointer_null) { //Make sure the parent exists and is not null
 	parentExists = instance_exists(parent);
 }
 
@@ -15,7 +15,7 @@ while (parentExists == false) and (ds_list_size(species.creatures) > 0) { //If t
 	ds_list_delete(species.creatures, ds_list_find_index(species.creatures, parent)); //Remove the parent from the species list
 	parent = ds_list_find_value(species.creatures, random(ds_list_size(species.creatures) - 0.5));
 	
-	if (parent != pointer_null) {
+	if (is_undefined(parent) == false) and (parent != pointer_null) and (instance_exists(parent) == true) {
 		parentExists = instance_exists(parent);
 	}
 
@@ -25,15 +25,17 @@ while (parentExists == false) and (ds_list_size(species.creatures) > 0) { //If t
 //Get a random creature for the parent
 
 //Create an egg
-egg = instance_create_layer(parent.x, parent.y, "EggLayer", obj_egg);
+if (ds_list_size(species.creatures) > 0) {
+	egg = instance_create_layer(parent.x, parent.y, "EggLayer", obj_egg);
 
-if (global.initialCreatureAmount >= 1) {
-	if (ds_list_size(species.creatures) >= global.creatureLimit) { //Upper limit (4x initial size) on population, kill the oldest creature
-		var toKill = ds_list_find_value(species.creatures, 0);
-		toKill.creatureHealth = 0;
+	if (global.initialCreatureAmount >= 1) {
+		if (ds_list_size(species.creatures) >= global.creatureLimit) { //Upper limit (4x initial size) on population, kill the oldest creature
+			var toKill = ds_list_find_value(species.creatures, 0);
+			toKill.creatureHealth = 0;
+		}
 	}
-}
 
-egg.parent = parent;
-egg.species = species;
-egg.alarm[0] = 5; //Initialize the egg
+	egg.parent = parent;
+	egg.species = species;
+	egg.alarm[0] = 5; //Initialize the egg
+}
